+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 
 ## [Unreleased]
+### Added
+- `read-file-attribute-view` method with corresponding records for default view
+  types.
+- `set-attribute` method.
 
 ## [0.1.5] — 2019-08-11
 ### Added